refactor: create library crate and clean up main crate
